Data.Matched.Cofree

Documentation

data MatchType

Input is finite foldable and what to do with the gaps between matched. Output is matched

Constructors

None

Parens

Bracket

Brace

newtype Matched a f b

Matches on inputs of type a with shape f and results of type b

 data Matched f a where
   Matched :: !MatchType -> f (Matched f a) -> Matched f a
   UnMatched :: !MatchType -> Fold (Matched f a) (f (Matched f a)) -> Matched f a

Constructors

Matched

Fields

runMatched :: CofreeT (FoldF f a) (Env MatchType) b

preMatched :: Functor f => (a -> b) -> Matched b f c -> Matched a f c